requirements.txt
django-crispy-forms>=1.8.1
django-import-export>=1.2.1.dev0
django-reversion>=3.0.4
django-formtools==2.1
future==0.15.2
httplib2==0.9.2
